Add files via upload
IPL Data Dashboard IPL is a popular cricket series played annually in India since 2007. So, we have generated a lot of data since then which helps in prediction of the future performances of teams and players. I have analysed the data which covers the various objectives- To analyse the player’s batting and bowling skills.

To compare the teams by their count number of match wins against each other.

To find out the popular playgrounds.

To show the toss distribution of team’s at different venues.

To find the number of boundaries by each team season-wise.

To show the players distribution in IPL from different countries.

Visulization of run rate.